requirements.txt
alembic==1.4.2
aliyun-python-sdk-core==2.13.32
aliyun-python-sdk-kms==2.14.0
amqp==2.5.2
billiard==3.5.0.5
blinker==1.4
celery==4.1.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
configparser==4.0.2
contextlib2==0.6.0.post1
crcmod==1.7
cryptography==2.9.2
enum34==1.1.10
Flask==0.10.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-MySQLdb==0.2.0
Flask-Script==2.0.6
Flask-Session==0.3.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==19.10.0
idna==2.9
importlib-metadata==1.6.0
ipaddress==1.0.23
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.10.0
kombu==4.6.8
Mako==1.1.2
MarkupSafe==1.1.1
mysqlclient==1.4.6
olefile==0.46
oss2==2.14.0
parse==1.15.0
pathlib2==2.3.5
Pillow==6.2.2
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.4.7
pycryptodomex==3.9.4
pyOpenSSL==19.1.0
python-alipay-sdk==1.4.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
qiniu==7.1.4
redis==3.2.0
requests==2.23.0
scandir==1.10.0
six==1.14.0
SQLAlchemy==1.3.16
urllib3==1.25.9
uWSGI==2.0.18
vine==1.3.0
Werkzeug==0.14.1
WTForms==2.2.1
xmltodict==0.11.0
zipp==1.2.0

